+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Next
 - Add ability to delete tracks from library
-- Add more items to Song menu
+- Add `Cmd+U` / `View > Toggle Queue` to show/hide the queue
+- Add more items to the Song menu
 - Add `Escape` shortcut for deselecting tracks
 - Add support selecting tracks via `Cmd+A`, `Shift+Up/Down`, `Alt+Up/Down`, `Shift+Alt+Up/Down`
 - Improve behavior of shift-selecting tracks